Rename module derivation -> address
