+++ packages/iov-ethereum/src/ethereumcodec.ts
@@ -19,8 +19,8 @@ import {
 import { ExtendedSecp256k1Signature, Keccak256, Secp256k1 } from "@iov/crypto";
 import { Encoding } from "@iov/encoding";
 
+import { isValidAddress, pubkeyToAddress, toChecksumAddress } from "./address";
 import { constants } from "./constants";
-import { isValidAddress, pubkeyToAddress, toChecksumAddress } from "./derivation";
 import { BlknumForkState, Eip155ChainId, getRecoveryParam } from "./encoding";
 import { Serialization } from "./serialization";
 import {
