+++ packages/iov-ethereum/types/index.d.ts
@@ -1,4 +1,4 @@
 export { EthereumConnection, EthereumConnectionOptions } from "./ethereumconnection";
 export { ethereumConnector } from "./ethereumconnector";
 export { ethereumCodec } from "./ethereumcodec";
-export { pubkeyToAddress } from "./derivation";
+export { pubkeyToAddress } from "./address";
